Remove the fhirmock file and instead update the basicmock to include the FhirEndpoint mock store. Address a comment in main.go about an error message.
