+++ endpointmanager/pkg/endpointmanager/mock/basicmock.go
@@ -8,12 +8,22 @@ import (
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager"
 )
 
+// BasicMockStore is a mocked store that includes both the HealthItProduct and FHIREndpoint stores
 type BasicMockStore struct {
 	HealthITProductData []*endpointmanager.HealthITProduct
+	FhirEndpointData    []*endpointmanager.FHIREndpoint
 	Store
 }
 
+// NewBasicMockHealthITProductStore returns mocked implementations of the functions
+// used in the HealthITProductStore store
 func NewBasicMockHealthITProductStore() endpointmanager.HealthITProductStore {
+	return newBasicMockStore()
+}
+
+// NewBasicMockFhirEndpointStore returns mocked implementations of the AddFhirEndpoint and GetFHIREndpointUsingURL functions
+// used in the FhirEndpoint store
+func NewBasicMockFhirEndpointStore() endpointmanager.FHIREndpointStore {
 	return newBasicMockStore()
 }
 
@@ -84,6 +94,46 @@ func newBasicMockStore() *BasicMockStore {
 		}
 
 		return nil
+	}
+
+	// FHIREndpointStore Functions
+
+	store.AddFHIREndpointFn = func(ctx context.Context, fhirEndpt *endpointmanager.FHIREndpoint) error {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		default:
+			// ok
+		}
+
+		for _, existingFhirEndpt := range store.FhirEndpointData {
+			if existingFhirEndpt.URL == fhirEndpt.URL {
+				return errors.New("FHIR Endpoint with that URL already exists")
+			}
+		}
+		// want to store a copy
+		newFhirEndpt := *fhirEndpt
+		newFhirEndpt.ID = len(store.FhirEndpointData) + 1
+		store.FhirEndpointData = append(store.FhirEndpointData, &newFhirEndpt)
+		return nil
+	}
+
+	store.GetFHIREndpointUsingURLFn = func(ctx context.Context, url string) (*endpointmanager.FHIREndpoint, error) {
+		select {
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		default:
+			// ok
+		}
+
+		for _, existingFhirEndpt := range store.FhirEndpointData {
+			if existingFhirEndpt.URL == url {
+				// want to return a copy
+				fhirEndpt := *existingFhirEndpt
+				return &fhirEndpt, nil
+			}
+		}
+		return nil, sql.ErrNoRows
 	}
 
 	return &store
